chore: refactor to follow standard pattern
